+++ fast/stages/0-bootstrap/organization.tf
@@ -50,14 +50,7 @@ locals {
     var.org_policies_config.constraints.allowed_policy_member_domains
   )
   drs_tag_name = "${var.organization.id}/${var.org_policies_config.tag_name}"
-  fast_custom_roles = [
-    "organization_admin_viewer",
-    "organization_iam_admin",
-    "service_project_network_admin",
-    "storage_viewer",
-    "tag_viewer",
-    "tenant_network_admin",
-  ]
+
   # intermediate values before we merge in what comes from the checklist
   _iam_principals = {
     for k, v in local.iam_principal_bindings : k => v.authoritative
@@ -101,9 +94,6 @@ locals {
     flatten(values(local._iam_principals)),
     keys(local._iam)
   ))
-  iam_roles_additive = distinct([
-    for k, v in local._iam_bindings_additive : v.role
-  ])
 }
 
 # TODO: add a check block to ensure our custom roles exist in the factory files