+++ fast/stages/1-resman/variables.tf
@@ -229,18 +229,6 @@ variable "locations" {
   nullable = false
 }
 
-variable "org_policy_tags" {
-  # tfdoc:variable:source 0-bootstrap
-  description = "Resource management tags for organization policy exceptions."
-  type = object({
-    key_id   = optional(string)
-    key_name = optional(string)
-    values   = optional(map(string), {})
-  })
-  nullable = false
-  default  = {}
-}
-
 variable "organization" {
   # tfdoc:variable:source 0-bootstrap
   description = "Organization details."